requirements.txt
Django==2.1.1
django-filter==2.0.0
django-model-utils==3.1.2
djangorestframework==3.8.2
pytz==2018.5
